fix warnings in examples

+++ docs/examples/c/utils/base64.c
@@ -58,14 +58,14 @@ int base64_decode(uint8_t *out, const char *in, int out_size)
     return dst - out;
 }
 
-char *base64_encode(char *out, int out_size, const uint8_t *in, int in_size)
+char *base64_encode(char *out, size_t out_size, const uint8_t *in, size_t in_size)
 {
     static const char b64[] =
         "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/";
     char *ret, *dst;
     unsigned i_bits = 0;
     int i_shift = 0;
-    int bytes_remaining = in_size;
+    size_t bytes_remaining = in_size;
 
     if (in_size >= UINT_MAX / 4 ||
         out_size < BASE64_SIZE(in_size))
